change main from agent to server
